backend/requirements.txt
asgiref==3.5.0
attrs==21.4.0
certifi==2021.10.8
charset-normalizer==2.0.11
Django==3.1.7
django-cleanup==6.0.0
django-cors-headers==3.7.0
djangorestframework==3.12.2
idna==3.3
iniconfig==1.1.1
packaging==21.3
Pillow==9.0.1
pluggy==1.0.0
psycopg2==2.9.3
py==1.11.0
pyparsing==3.0.7
pytest==7.0.0
pytz==2021.3
requests==2.27.1
sqlparse==0.4.2
tomli==2.0.1
urllib3==1.26.8
